src/lib/utils/windows.rs:migrate windows_process_killer from src/lib/ws.rs
